requirements.txt
fastapi
uvicorn
requests
deep-translator
pydantic
google-auth